Enable dictionary encoding for FIXED.
This uses the existing dictionary support introduced for int96. Encoding
and ParquetProperties have been updated to use the dictionary supporting
classes, when requested for write or present during read. This also
fixes a bug in the fixed dictionary values writer, where the length was
hard-coded for int96, 12 bytes.

Conflicts:
	parquet-column/src/main/java/parquet/column/ParquetProperties.java

Changes applied to ColumnWriterImpl instead of ParquetProperties. The
change to separate out ParquetProperties was not backported.