CLOUDERA-BUILD. Remove use of TBinaryProtocol#setReadLength.
This is no longer supported in thrift 0.9.2 and was only used
defensively. The reason to remove it now is to avoid linker errors when
the wrong version of thrift is found in the classpath.

Upstream will probably add a dynamic call to this method when it is
present, but this depends on 0.9.2 so it is not present.