PARQUET-217 Use simpler heuristic in MemoryManager
We found that the heuristic of throwing when:
```
minMemoryAllocation > 0 && newSize/maxColCount < minMemoryAllocation
```
in MemoryManager is not really valid when you have many (3k +) columns, due to the division by the number of columns.
This check throws immediately when writing a single file with a 3GB heap and > 3K columns.

This PR introduces a simpler heuristic, which is a min scale, and we throw when the MemoryManager's scale gets too small. By default I chose 25%, but I'm happy to change that to something else.

For backwards compatibility I've left the original check in, but it's not executed by default anymore, to get this behavior the min chunk size will have to be set in the hadoop configuration. I'm also open to removing it entirely if we don't think we need it anymore.

What do you think?
@danielcweeks @rdblue @dongche @julienledem

Author: Alex Levenson <alexlevenson@twitter.com>

Closes #143 from isnotinvain/alexlevenson/mem-manager-heuristic and squashes the following commits:

acda66f [Alex Levenson] Add units to exception
10237c6 [Alex Levenson] Decouple DEFAULT_MIN_MEMORY_ALLOCATION from DEFAULT_PAGE_SIZE
29c9881 [Alex Levenson] Use an absolute minimum on rowgroup size, only apply when scale < 1
8877125 [Alex Levenson] Merge branch 'master' into alexlevenson/mem-manager-heuristic
e5117a0 [Alex Levenson] Merge branch 'master' into alexlevenson/mem-manager-heuristic
6ee5f46 [Alex Levenson] Use simpler heuristic in MemoryManager