PARQUET-340: MemoryManager: max memory can be truncated
Using float will cause the max heap limit to be limited to 2147483647
due to math.round(float) if used with a large heap. This should be a double
precision to prevent rounding to an int32 before storing into a long.

Author: Chris Bannister <c.bannister@gmail.com>

Closes #246 from Zariel/default-mem-truncated and squashes the following commits:

bf375f6 [Chris Bannister] MemoryManager: ensure max memory is not truncated