PARQUET-335: Remove Avro check for MAP_KEY_VALUE.
This is not required by the map type spec. This does not affect data
written by the Avro object model because this bug is in the conversion
from a Parquet schema to an Avro schema. Files written with parquet-avro
do not convert the underlying schema because they use the Avro schema.

Author: Ryan Blue <blue@apache.org>

Closes #241 from rdblue/PARQUET-335-remove-key-value-check and squashes the following commits:

1fd9541 [Ryan Blue] PARQUET-335: Test that MAP_KEY_VALUE is not required.
247cc76 [Ryan Blue] PARQUET-335: Remove Avro check for MAP_KEY_VALUE.